refactor(palette): split file listing strategies and extract fuzzy filter
Each strategy (rg, git, walkDir) is a standalone function returning
a file list. listDirFiles tries them in order and returns the first
success. fuzzyFilterFiles is now a standalone function in filelist.go,
keeping the dialog widget focused on UI.
