+++ internal/ui/filelist.go
@@ -4,6 +4,7 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
+	"sort"
 	"strings"
 )
 
@@ -12,28 +13,28 @@ const walkDirMaxFiles = 100000
 func listWorkspaceFiles(workDirs []string) []paletteFile {
 	var files []paletteFile
 	multiRoot := len(workDirs) > 1
-	_, hasRg := exec.LookPath("rg")
-	_, hasGit := exec.LookPath("git")
 	for _, workDir := range workDirs {
 		prefix := ""
 		if multiRoot {
 			prefix = filepath.Base(workDir) + string(filepath.Separator)
 		}
-		if hasRg == nil {
-			if f, ok := listFilesCmdLines(workDir, prefix, "rg", "--files"); ok {
-				files = append(files, f...)
-				continue
-			}
-		}
-		if hasGit == nil {
-			if f, ok := listFilesCmdLines(workDir, prefix, "git", "ls-files", "--cached", "--others", "--exclude-standard"); ok {
-				files = append(files, f...)
-				continue
-			}
-		}
-		files = append(files, listFilesWalkDir(workDir, prefix)...)
+		files = append(files, listDirFiles(workDir, prefix)...)
 	}
 	return files
+}
+
+func listDirFiles(workDir, prefix string) []paletteFile {
+	if _, err := exec.LookPath("rg"); err == nil {
+		if files, ok := listFilesCmdLines(workDir, prefix, "rg", "--files"); ok {
+			return files
+		}
+	}
+	if _, err := exec.LookPath("git"); err == nil {
+		if files, ok := listFilesCmdLines(workDir, prefix, "git", "ls-files", "--cached", "--others", "--exclude-standard"); ok {
+			return files
+		}
+	}
+	return listFilesWalkDir(workDir, prefix)
 }
 
 func listFilesCmdLines(workDir, prefix string, name string, args ...string) ([]paletteFile, bool) {
@@ -80,4 +81,58 @@ func listFilesWalkDir(workDir, prefix string) []paletteFile {
 		return nil
 	})
 	return files
+}
+
+func fileDetail(f string) string {
+	dir := filepath.Dir(f)
+	if dir == "." {
+		return ""
+	}
+	return dir
+}
+
+func fuzzyFilterFiles(files []paletteFile, query string, maxResults int) []PaletteItem {
+	if query == "" {
+		items := make([]PaletteItem, 0, min(len(files), maxResults))
+		for _, f := range files {
+			items = append(items, PaletteItem{
+				Label:  filepath.Base(f.Rel),
+				Detail: fileDetail(f.Rel),
+				ID:     f.Abs,
+			})
+			if len(items) >= maxResults {
+				break
+			}
+		}
+		return items
+	}
+
+	type scored struct {
+		item  PaletteItem
+		score int
+	}
+	var matches []scored
+	for _, f := range files {
+		if ok, score := fuzzyMatch(query, f.Rel); ok {
+			matches = append(matches, scored{
+				item: PaletteItem{
+					Label:  filepath.Base(f.Rel),
+					Detail: fileDetail(f.Rel),
+					ID:     f.Abs,
+				},
+				score: score,
+			})
+		}
+	}
+	sort.Slice(matches, func(i, j int) bool {
+		return matches[i].score > matches[j].score
+	})
+	items := make([]PaletteItem, 0, min(len(matches), maxResults))
+	for _, m := range matches {
+		items = append(items, m.item)
+		if len(items) >= maxResults {
+			break
+		}
+	}
+	return items
 }
