+++ internal/ui/selectdialog_widget.go
@@ -2,7 +2,6 @@ package ui
 
 import (
 	"fmt"
-	"path/filepath"
 	"sort"
 	"strconv"
 	"strings"
@@ -605,55 +604,8 @@ func (p *SelectDialogWidget) filterCommands(query string) {
 	p.scrollOffset = 0
 }
 
-func fileDetail(f string) string {
-	dir := filepath.Dir(f)
-	if dir == "." {
-		return ""
-	}
-	return dir
-}
-
 func (p *SelectDialogWidget) filterFiles(query string) {
-	p.Items = nil
-	if query == "" {
-		for _, f := range p.files {
-			p.Items = append(p.Items, PaletteItem{
-				Label:  filepath.Base(f.Rel),
-				Detail: fileDetail(f.Rel),
-				ID:     f.Abs,
-			})
-			if len(p.Items) >= 100 {
-				break
-			}
-		}
-	} else {
-		type scored struct {
-			item  PaletteItem
-			score int
-		}
-		var matches []scored
-		for _, f := range p.files {
-			if ok, score := fuzzyMatch(query, f.Rel); ok {
-				matches = append(matches, scored{
-					item: PaletteItem{
-						Label:  filepath.Base(f.Rel),
-						Detail: fileDetail(f.Rel),
-						ID:     f.Abs,
-					},
-					score: score,
-				})
-			}
-		}
-		sort.Slice(matches, func(i, j int) bool {
-			return matches[i].score > matches[j].score
-		})
-		for _, m := range matches {
-			p.Items = append(p.Items, m.item)
-			if len(p.Items) >= 100 {
-				break
-			}
-		}
-	}
+	p.Items = fuzzyFilterFiles(p.files, query, 100)
 	p.Selected = 0
 	p.scrollOffset = 0
 }
